Move debug buttons to secret debug menu
