+++ app/src/main/res/layout/activity_debug.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <Button
+        android:id="@+id/chainExplorerButton"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/chain_explorer" />
+
+    <Button
+        android:id="@+id/resetDatabaseButton"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="8dp"
+        android:text="@string/reset_database_button"
+        app:layout_constraintTop_toBottomOf="@+id/chainExplorerButton"
+        tools:layout_editor_absoluteX="0dp" />
+
+    <Button
+    android:id="@+id/debugFingerprintButton"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:text="Activate fingerprint activity"
+    android:layout_marginTop="8dp"
+    app:layout_constraintTop_toBottomOf="@+id/resetDatabaseButton"
+        />
+
+</android.support.constraint.ConstraintLayout>
