+++ app/src/main/res/values/strings.xml
@@ -34,5 +34,7 @@
     <string name="bluetooth_explanation">First pair with a device via the Android bluetooth manager. After the two devices are paired press on the desired device to initiate a connection.</string>
     <string name="bluetooth_button">Bluetooth</string>
     <string name="FingerprintDeclinedText">Fingerprint declined. Please try again</string>
+    <string name="choose_device_to_connect">Choose device to connect</string>
+    <string name="confirmation_button_text">Confirm</string>
 
 </resources>
